Streaming working, with some ugly code!

+++ src/main/cpp/Channel.cpp
@@ -7,7 +7,7 @@
 using namespace IMS;
 
 JNIEXPORT jint JNICALL Java_org_lsst_ccs_daq_imageapi_DAQSourceChannel_read
-(JNIEnv *env, jobject obj, jlong source_, jobject buffer, jint position, jint offset, jint length) {
+(JNIEnv *env, jobject obj, jlong source_, jobject buffer, jint position, jlong offset, jint length) {
     Source* source = (Source*) source_;
     char *buf = (char*) env->GetDirectBufferAddress(buffer);
     return source->read(buf + position, length, offset);
